Replaced score type with Int
See issue https://github.com/JetBrains-Research/bioinf-commons/issues/2

+++ src/test/kotlin/org/jetbrains/bio/big/BedEntryTest.kt
@@ -317,7 +317,7 @@ class BedEntryTest {
         val actualFields = (0 until 14).map { BED_ENTRY_12_P_2.getField(it, fieldsNumber, extraFieldsNumber) }
         val realExtraFieldsNumber = extraFieldsNumber ?: 2
         val expectedFields = listOf<Any?>(
-            "chr1", 10, 30, "be", 5.toShort(), '+', 15, 25, Color(15, 16, 17).rgb,
+            "chr1", 10, 30, "be", 5, '+', 15, 25, Color(15, 16, 17).rgb,
             2, intArrayOf(4, 5), intArrayOf(11, 20)
         ).slice(0 until fieldsNumber).toMutableList()
         expectedFields.addAll(listOf("val1", "4.55").slice(0 until realExtraFieldsNumber))
